[flang][runtime] Allow different memmov function in assign
